% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_counties}
\alias{get_counties}
\title{List counties where a species is found}
\usage{
get_counties(species)
}
\arguments{
\item{species}{Scientific name of the species to query}
}
\value{
A tibble of state & county-level occurrence data
}
\description{
List counties where a species is found
}
\examples{
get_counties("Ursus horribilis")
}
\seealso{
\link{get_species}
}

