% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_species}
\alias{get_species}
\title{List species found in a county}
\usage{
get_species(fips = "", state = "", county = "")
}
\arguments{
\item{fips}{FIPS code for the county to query}
}
\value{
A tibble of state & county-level occurrence data
}
\description{
List species found in a county
}
\examples{
get_species(state = "VA", county = "Clarke")
}
\seealso{
\link{get_counties}
}

